+++ Carbon.IIS/Carbon.IIS.psm1
@@ -24,6 +24,18 @@ $moduleRoot = $PSScriptRoot
 Import-Module -Name (Join-Path -Path $moduleRoot -ChildPath 'PSModules\Carbon.Core' -Resolve) `
               -Function @('Add-CTypeData')
 
+enum CIisNumaNodeAffinityMode
+{
+    Soft = [UInt32]0
+    Hard = [UInt32]1
+}
+
+enum CIisNumaNodeAssignment
+{
+    MostAvailableMemory = [UInt32]0
+    WindowsScheduling = [UInt32]1
+}
+
 function Test-MSWebAdministrationLoaded
 {
     $serverMgrType =